// src/types/api-response.ts
// laravel's pagination response
export interface ApiResponse<T> {
  status: string;
  message: string;
  data: T;
}

export interface PaginationLink {
  url: string | null;
  label: string;
  active: boolean;
}

export interface Pagination<T> {
  current_page: number;
  data: T[];
  first_page_url: string;
  from: number | null;
  last_page: number;
  last_page_url: string;
  links: PaginationLink[];
  next_page_url: string | null;
  path: string;
  per_page: number;
  prev_page_url: string | null;
  to: number | null;
  total: number;
}

export type PaginatedResponse<T> = ApiResponse<Pagination<T>>;